+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Asset Indexes Extra for Craft 5
 
+## Unreleased
+
+- Fixed a bug on plugin installation for MySQL on index creation.
+
 ## 2.0.0-alpha.1 - 2024-06-28
 
 - Initial Release